+++ src/CGCLib/CgcClientHandler.h
@@ -102,7 +102,7 @@ namespace cgc
 		virtual void doUnRegisterAllSink(cgc::bigint nRoomId) = 0;
 		virtual void doUnRegisterAllSink(void) = 0;
 		virtual bool doIsRegisterSink(cgc::bigint nRoomId, cgc::bigint nDestId) const = 0;
-		virtual bool doSendRtpData(cgc::bigint nRoomId,const unsigned char* pData,cgc::uint16 nSize,cgc::uint32 nTimestamp=0,cgc::uint8 nDataType=SOTP_RTP_NAK_DATA_AUDIO,cgc::uint8 nNAKType=SOTP_RTP_NAK_REQUEST_1) = 0;
+		virtual bool doSendRtpData(cgc::bigint nRoomId,const unsigned char* pData,cgc::uint32 nSize,cgc::uint32 nTimestamp=0,cgc::uint8 nDataType=SOTP_RTP_NAK_DATA_AUDIO,cgc::uint8 nNAKType=SOTP_RTP_NAK_REQUEST_1) = 0;
 
 		// thread
 		virtual void doSetCIDTResends(unsigned short timeoutResends=2, unsigned short timeoutSeconds=4) = 0;
